+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @triplit/svelte
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies [c3fb0be]
+  - @triplit/client@0.3.73
+
+## 0.1.44
+
+### Patch Changes
+
+- Updated dependencies [607e020]
+  - @triplit/client@0.3.72
+
+## 0.1.43
+
+### Patch Changes
+
+- @triplit/client@0.3.71
+
 ## 0.1.42
 
 ### Patch Changes