+++ templates/chat-template/CHANGELOG.md
@@ -4,6 +4,33 @@
 
 ### Patch Changes
 
+- Updated dependencies [c3fb0be]
+- Updated dependencies [c3fb0be]
+  - @triplit/client@0.3.73
+  - @triplit/db@0.3.56
+  - @triplit/react@0.3.77
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [607e020]
+  - @triplit/client@0.3.72
+  - @triplit/react@0.3.76
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [b65160e]
+  - @triplit/db@0.3.55
+  - @triplit/client@0.3.71
+  - @triplit/react@0.3.75
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [55ae69b]
 - Updated dependencies [2d2b360]
 - Updated dependencies [a8c6f75]